Made that commands ain't hardcoded to overworld
This should be a good fix, it will not handle to create dimensions tho, there is other resources better than the mod can provide, it would require some restructuring that is not worth the hassle, the main problem is that commands are defaulted to overworld not target dimension under config.
Also provided some clear messages in case someone is trying to do it incorrectly, for my case it worked just plain fine.